third_party/WebKit/LayoutTests/inspector/tracing/tracing-timeline-load-expected.txt

Tests tracing based Timeline save/load functionality.


Running: testNormal
TimelineLoaderClient.loadingStarted()
TimelineLoaderClient.loadingProgress()
TimelineLoaderClient.loadingComplete(true)
Saved data is equal to restored data: true

Running: testJSONObjectFormat
TimelineLoaderClient.loadingStarted()
TimelineLoaderClient.loadingProgress()
TimelineLoaderClient.loadingComplete(true)
Saved data is equal to restored data: true

Running: testJSONObjectFormatWithMetadata
TimelineLoaderClient.loadingStarted()
TimelineLoaderClient.loadingProgress()
TimelineLoaderClient.loadingComplete(true)
Saved data is equal to restored data: true

Running: testBroken
TimelineLoaderClient.loadingStarted()
TimelineLoaderClient.loadingComplete(true)
Model is empty: true

Running: testMalformedJSON
TimelineLoaderClient.loadingComplete(false)
Model is empty: true
